+++ packages/solverpy/src/solverpy/solver/plugins/db/__init__.py
@@ -28,6 +28,7 @@ class solverpy.solver.plugins.db.errors.Errors extends solverpy.solver.plugins.d
 
 from .bid import Bid
 from .sid import Sid
+from .eprovesid import EProverSid
 
-__all__ = ["Bid", "Sid"]
+__all__ = ["Bid", "Sid", "EProverSid"]